Let CI jobs run on all PRs, not just those rooted off master.
I want to use stack diffs (https://github.com/ezyang/ghstack/) to
submit PRs to Cabal.  Stack diffs, simply by virtue of what they
are, are PRs that are targeted to branches that are not master (they
target an appropriate base that lets GitHub render only the change
in question, even if there are other changes below it.)  However,
I don't get CI jobs on those stack diffs because the current
pull_request branch selector is too restrictive.  This PR makes
it less restrictive.

This PR is submitted using ghstack to validate that this indeed
works.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: b135ea67848f873fc6b50b950405859587ae8d89
Pull Request resolved: https://github.com/haskell/cabal/pull/7308