Canonicalize path in TargetSelector to fix No Targets bug on Windows
On line 1752 in TargetSelector a call to getCurrentDirectory is not
canonicalized which prevents cabal-install from finding the primary
package on Windows when no targets are specified.

This happens because on line 1772 the cwd is compared as a string with
pinfo path which is canoncialized so we get:

       `c:\myProject\` /= `C:\myProject\`.

This only happens under certain shells but it does completely break
cabal functionality in emacs' haskell-mode (on Windows) as it relies
on automatic package detection.